[HttpClient] Fix codereview issues
