+++ httpclient-backend/src/main/scala/sttp/client/httpclient/HttpClientFutureBackend.scala
@@ -5,17 +5,12 @@ import java.net.http.HttpRequest.BodyPublisher
 import java.net.http.{HttpClient, HttpRequest}
 import java.nio.ByteBuffer
 import java.util
-import java.util.concurrent.LinkedBlockingQueue
-import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
-import java.util.function.UnaryOperator
-import java.util.stream.{Collector, Collectors}
 import java.util.zip.{GZIPInputStream, InflaterInputStream}
 
-import org.reactivestreams.{Publisher, Subscriber, Subscription}
+import org.reactivestreams.Publisher
 import sttp.capabilities.WebSockets
 import sttp.client.httpclient.HttpClientBackend.EncodingHandler
-import sttp.client.httpclient.HttpClientFutureBackend.FutureEncodingHandler
-import sttp.client.httpclient.InputStreamSubscriber.{Completed, Error, Message, NextItem, nextItemMsg}
+import sttp.client.httpclient.HttpClientFutureBackend.InputStreamEncodingHandler
 import sttp.client.internal.ws.{FutureSimpleQueue, SimpleQueue}
 import sttp.client.internal.{NoStreams, emptyInputStream}
 import sttp.client.testing.SttpBackendStub
@@ -29,7 +24,7 @@ class HttpClientFutureBackend private (
     client: HttpClient,
     closeClient: Boolean,
     customizeRequest: HttpRequest => HttpRequest,
-    customEncodingHandler: FutureEncodingHandler
+    customEncodingHandler: InputStreamEncodingHandler
 )(implicit ec: ExecutionContext)
     extends HttpClientAsyncBackend[Future, Nothing, WebSockets, InputStream](
       client,
@@ -78,13 +73,13 @@ class HttpClientFutureBackend private (
 }
 
 object HttpClientFutureBackend {
-  type FutureEncodingHandler = EncodingHandler[InputStream]
+  type InputStreamEncodingHandler = EncodingHandler[InputStream]
 
   private def apply(
       client: HttpClient,
       closeClient: Boolean,
       customizeRequest: HttpRequest => HttpRequest,
-      customEncodingHandler: FutureEncodingHandler
+      customEncodingHandler: InputStreamEncodingHandler
   )(implicit ec: ExecutionContext): SttpBackend[Future, WebSockets] =
     new FollowRedirectsBackend(
       new HttpClientFutureBackend(client, closeClient, customizeRequest, customEncodingHandler)
@@ -93,7 +88,7 @@ object HttpClientFutureBackend {
   def apply(
       options: SttpBackendOptions = SttpBackendOptions.Default,
       customizeRequest: HttpRequest => HttpRequest = identity,
-      customEncodingHandler: FutureEncodingHandler = PartialFunction.empty
+      customEncodingHandler: InputStreamEncodingHandler = PartialFunction.empty
   )(implicit ec: ExecutionContext = ExecutionContext.global): SttpBackend[Future, WebSockets] =
     HttpClientFutureBackend(
       HttpClientBackend.defaultClient(options),
@@ -105,7 +100,7 @@ object HttpClientFutureBackend {
   def usingClient(
       client: HttpClient,
       customizeRequest: HttpRequest => HttpRequest = identity,
-      customEncodingHandler: FutureEncodingHandler = PartialFunction.empty
+      customEncodingHandler: InputStreamEncodingHandler = PartialFunction.empty
   )(implicit ec: ExecutionContext = ExecutionContext.global): SttpBackend[Future, WebSockets] =
     HttpClientFutureBackend(client, closeClient = false, customizeRequest, customEncodingHandler)
 
@@ -116,101 +111,4 @@ object HttpClientFutureBackend {
     */
   def stub(implicit ec: ExecutionContext = ExecutionContext.global): SttpBackendStub[Future, WebSockets] =
     SttpBackendStub(new FutureMonad())
-}
-
-// based on org.asynchttpclient.request.body.generator.ReactiveStreamsBodyGenerator.SimpleSubscriber
-private[httpclient] class InputStreamSubscriber extends Subscriber[java.util.List[ByteBuffer]] {
-  // a pair of values: (is cancelled, current subscription)
-  private val subscription = new AtomicReference[(Boolean, Subscription)]((false, null))
-  private val chunks = new LinkedBlockingQueue[Message]()
-
-  val inputStream: InputStream = new InputStream {
-    private val exhausted = new AtomicBoolean(false)
-    private val currentBuffer: AtomicReference[Option[ByteBuffer]] = new AtomicReference[Option[ByteBuffer]](None)
-
-    override def read(): Int = {
-      if (exhausted.get()) {
-        -1
-      } else {
-        val byteRead = currentBuffer.get() match {
-          case Some(buffer) if buffer.hasRemaining =>
-            buffer.get() & 0xff
-          case _ =>
-            chunks.take() match {
-              case NextItem(buffer) =>
-                currentBuffer.set(Some(buffer))
-                buffer.get() & 0xff
-              case Error(ex) =>
-                throw ex
-              case Completed() =>
-                exhausted.set(true)
-                -1
-            }
-        }
-        byteRead
-      }
-    }
-  }
-
-  override def onSubscribe(s: Subscription): Unit = {
-    assert(s != null)
-
-    // The following can be safely run multiple times, as cancel() is idempotent
-    val result = subscription.updateAndGet(new UnaryOperator[(Boolean, Subscription)] {
-      override def apply(current: (Boolean, Subscription)): (Boolean, Subscription) = {
-        // If someone has made a mistake and added this Subscriber multiple times, let's handle it gracefully
-        if (current._2 != null) {
-          current._2.cancel() // Cancel the additional subscription
-        }
-
-        if (current._1) { // already cancelled
-          s.cancel()
-          (true, null)
-        } else { // happy path
-          (false, s)
-        }
-      }
-    })
-
-    if (result._2 != null) {
-      result._2.request(Long.MaxValue) // not cancelled, we can request data
-    }
-  }
-
-  private val toListCollector: Collector[Message, _, util.List[Message]] = Collectors.toList()
-  override def onNext(b: java.util.List[ByteBuffer]): Unit = {
-    assert(b != null)
-    chunks.addAll(b.stream().map(nextItemMsg(_)).collect(toListCollector))
-  }
-
-  override def onError(t: Throwable): Unit = {
-    assert(t != null)
-    chunks.add(Error(t))
-  }
-
-  override def onComplete(): Unit = {
-    chunks.add(Completed())
-  }
-
-  def cancel(): Unit = {
-    // subscription.cancel is idempotent:
-    // https://github.com/reactive-streams/reactive-streams-jvm/blob/v1.0.3/README.md#specification
-    // so the following can be safely retried
-    subscription.updateAndGet(new UnaryOperator[(Boolean, Subscription)] {
-      override def apply(current: (Boolean, Subscription)): (Boolean, Subscription) = {
-        if (current._2 != null) current._2.cancel()
-        (true, null)
-      }
-    })
-  }
-}
-
-object InputStreamSubscriber {
-  sealed trait Message
-
-  case class NextItem(payload: ByteBuffer) extends Message
-  case class Error(ex: Throwable) extends Message
-  case class Completed() extends Message
-
-  def nextItemMsg(payload: ByteBuffer): Message = NextItem(payload)
 }
